Updated seeders to use firstOrCreate method
This change updates the database seeders for Subregion, Country, State, City, and Region to use the firstOrCreate method instead of create. This prevents duplicate entries from being inserted into the database, as firstOrCreate will only create a new record if one does not already exist with the specified unique attributes.